requirements.txt
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
packaging==24.1
psutil==6.0.0
pyyaml==6.0.1
requests==2.32.3
transformers==4.43.3
accelerate==0.33.0
flask==3.0.3
langid==1.1.6